test: 拆分 Agent 选中验收并等待 Codex 首屏
